refactor!: encapsulate codes into class ToolPanel
BREAKING CHANGE: migrate codes related to tool panel to the class
ToolPanel and the file tool-panel.css.

There is only one public API for class ToolPanel, which is the property
mainPanelDiv, which is the HTMLDivElement of itself.

A DoraemonCanvas instance should be responsible in creating a ToolPanel
instance in the constructor and have a has-a relationship to it.
Normally, it should also be the only channel through which outside can
get access to the public API of the ToolPanel instance.